==> 2_basic_process_data/2_processed_raw_data.ipynb <==
# Limpieza data raw "correcta"

- Se toma la data raw "correcta", la data preprocessed, que como dice el nombre es la data raw correguido valores nulos por fallas de conexión con PI, setear segundos a cero, etc.

- **Limpiezas hechas**:
    - Suavizar
    - Timegap
    - Pivot data
    - Eliminar PGP

    
- **Cambios con respecto a la limpieza ORIGINAL DEL RECOMENDADOR**:
    - Se suaviza solo con 2 periodos como ventanas en lugar de 5
    - Se utiliza el valor promedio del timegap para suavizar en lugar de un timegap dinámico (simplicar). Punto de observación al inicio del proceso
    - Se guarda en una columna el valor actual de tag y en otra columna el tag desplazado (CREACIÓN NUEVA NO EXISTENTE ANTES)

-------
**DATA**:
- INPUT: "data_raw_preprocessed.pkl"
- OUTPUT: "data_raw_processed.pkl"

## Root folder and read env variables

import os
# fix root path to save outputs
actual_path = os.path.abspath(os.getcwd())
list_root_path = actual_path.split('\\')[:-1]
root_path = '\\'.join(list_root_path)
os.chdir(root_path)
print('root path: ', root_path)

import os
from dotenv import load_dotenv, find_dotenv # package used in jupyter notebook to read the variables in file .env

""" get env variable from .env """
load_dotenv(find_dotenv())

""" Read env variables and save it as python variable """
PROJECT_GCP = os.environ.get("PROJECT_GCP", "")

## RUN

import pandas as pd
import numpy as np
import datetime as dt
import json
import pickle
from sklearn.pipeline import Pipeline
import sys
import os
import matplotlib.pyplot as plt
import gcsfs

import warnings
warnings.filterwarnings("ignore")

from sklearn.base import BaseEstimator, TransformerMixin

### 0. Funciones Auxiliares

# funciones auxiliares
def load_all_parameterstags_tagclassification(model_name):
    """
    Read a dictionary with all parameters filtered by model (d0eop, d1d2, d2, etc) located in TagClassification 
    """
    path_json = 'config/params.json'
    with open("{path}".format(path=path_json)) as json_file:
        tag_classification_pars = json.load(json_file)

    return tag_classification_pars[model_name]

# define name of model - transversal model for this example
general_params_models = 'blanqueo_santafe_all'

### 1. Leer data raw - datalake
### 1. Read data raw datalake - preprocessed
- Data get in the previous notebook
- Data without nulls - filled in previous step for problems in upload data, no conextion PI-datalake, etc

path_raw_data = f'artifacts/data/data_raw_preprocessed.pkl'
basic_preprocessed_data = pd.read_pickle(path_raw_data)
basic_preprocessed_data.head(3)

print('primera fecha: ', basic_preprocessed_data['datetime'].min())
print('ultima fecha: ', basic_preprocessed_data['datetime'].max())

# validate number of nulls
df_aux = basic_preprocessed_data.pivot(index='datetime', columns='Tag', values='PV')
df_aux.isnull().sum().sum()

## Continuar los pasos de limpieza después de obtener data raw correcta sin problemas de valores nulos

### 5. Suavizar Data
En el json está el parámetro de periodos que se utilizan para suavizar

class smooth_data(BaseEstimator,TransformerMixin):
    '''
    Suavizar la data. 
    - Actualmente suavizar toda la data (excepto los tags de tiempo de retención) de acuerdo a un parámetro fijo
    - Se suaviza cuando está toda la data con valores y NO HAY NULOS, así evito que se expandan los nulos al aplicar el rolling
    
    - Obs: recibe la data raw, pivotear, suaviza, vuelve
    '''
    
    def __init__(self, freq_rolling_smooth_data, timegap_tag_list):
        super(smooth_data,self).__init__()
        self.freq_rolling_smooth_data = freq_rolling_smooth_data # freq con la que se va suavizar
        self.timegap_tag_list = timegap_tag_list # listado de tags de tiempos de retención

    def fit(self,DataFrame):
        return self

    def transform(self,DataFrame):
        print('\napplying smooth_data...')
        
        # pivotear data
        print('Tamaño data raw: ', DataFrame.shape)
        DataFrame = DataFrame.pivot(index='datetime', columns='Tag', values='PV')

        # definir tags que se van a suavizar -> todos excepto los tag de timegap (esos datos tienen su propio tratamiento en otra función)
        list_tags_to_smooth = list(set(DataFrame.columns.tolist()) - set(self.timegap_tag_list ))
        
        # suavizar la data y borrar nulos generados
        DataFrame[list_tags_to_smooth] = DataFrame[list_tags_to_smooth].rolling(self.freq_rolling_smooth_data).mean()
        DataFrame = DataFrame.dropna()
        print('Cantidad de datos a borrarse por suavizar la data(primeros valores): ', (self.freq_rolling_smooth_data - 1) * DataFrame.shape[1])
        
        
        # unpivot de la data porque el timegap dinámico está armado en base a la estructura de la data raw
        DataFrame = DataFrame.reset_index()
        DataFrame = pd.melt(DataFrame, id_vars = ['datetime'])
        DataFrame.rename(columns = {'value': 'PV'}, inplace = True)
        print('Tamaño data luego de suavizar: ', DataFrame.shape)
        
        return DataFrame

# IMPORTANTE
# ventana filtro ma - AHORA ES DE 2 PERIODOS COMO VENTANA (en lugar de 5 periodos como tiene la limpiezaz original)
selected_model_name = 'blanqueo_santafe_all'
tag_classification_pars = load_all_parameterstags_tagclassification(model_name = general_params_models)
moving_average_all_tags = tag_classification_pars['cleaning_pars']['moving_average_all_data']
moving_average_all_tags

# read master tag table
path_MaestroEtapas = 'config/MaestroEtapasGlobal.xlsx'
MaestroEtapas = pd.read_excel(path_MaestroEtapas)

# list of tags wuith values of timegap - this tags has its own process of cleaning
timegaptag_list = list(MaestroEtapas.dropna()['TAG_TIEMPO_RESIDENCIA'])

# instancia de la clase
smooter = smooth_data(freq_rolling_smooth_data = moving_average_all_tags, 
                      timegap_tag_list = timegaptag_list
                     )

# transform
preprocessed_data = smooter.transform(basic_preprocessed_data)

### 6. TimeGap PROMEDIO
- En la limpieza original se utiliza los timegap dinámicos (se utilizan tiempos de residencia como función de la producción y el tamaño del tanque). Esto tiene diferentes problemas al generar muchos nulos

- En este ejemplo sencillo, se obviará los timegap dinámicos y se utilizará un TIMEGAP PROMEDIO (donde este valor promedio se obtuvo en combinación con los tiempos de retención promedio que manejan los ingenieros de procesos y validado a través de los tags de tiempo de renteción obteniendo el promedio de cada uno de lo valores de tiempo de residencia)

- Al utilizar simplemente un desplazamiento promedio no se generan nulos y es relativamente cercano a la realidad (por lo menos cuando el proceso es estable el cual cual debería ser la mayor parte del tiempo)

- Para hacer los desplazamiento de la data me voy a parar AL INICIO DEL PROCESO en la etapa D0 (en una primera instancia no se va a considerar la etapa ácida)

#### 6.1 Auxiliar. Crear diferentes tags de producción para ser desplazados
- Originalmente existen los tags de producción a la entrada de blanqueo (etapa A), entrada de eop y bypass.
- Se van a crear tags de producción para la etapa D0, D1 y P (simplemente tomar la producción del blanqueo a la entrada y crear las variables de producción a la entrada de las etapas D0, D1 y P, las que posteriormente se desplazaran de acuerdo al tiempo de residencia 

def crear_tags_produccion_cada_etapa(df):
    '''
    Crear tag de producción en las etapas que no existe directamente el tag de producción: D0, D1, P
    '''
    # pivotear data
    df_pivot = df.pivot(index='datetime', columns='Tag', values='PV')

    """
    ######### DISCOVERY #########
    ### DISCOVERY - DIFERENCIA ENTRE LOS VALORES DE (PRODUCCIÓN A LA ENTRADA DE BLANQUEO) VS (PRODUCCIÓN ENTRADA E0P + BYPASS)
    ### DISCOVERY - SE OBSERVA QUE LA DIFERENCIA ENTRE AMBAS VOLARES RONDA ENTRE 20 A 70 ADT

    list_tags_produccion = ['240FI020A.PNT', '240FI020B.PNT', '240FI108A.PNT']
    df_pivot['prod_eop_bypass'] = df_pivot['240FI020B.PNT'] + df_pivot['240FI108A.PNT']
    df_pivot['diff_prod_d0_eopbypass'] = df_pivot['240FI020A.PNT'] - df_pivot['prod_eop_bypass']
    print('\nDescribe diferencia')
    print(df_pivot['diff_prod_d0_eopbypass'].describe(percentiles = [0.05, 0.1, 0.25, 0.5, 0.75, 0.90, 0.95]))
    print('\nHistograma diferencia')
    df_pivot['diff_prod_d0_eopbypass'].hist()
    """

    # crear tags de producción - aún no se desplaza la data - en el maestro tag está definido el tag con el nombre que se va a
    # crear para desplazar los valores de producción a su torre correspondiente
    df_pivot['calc_prod_d0'] = df_pivot['240FI020A.PNT']
    df_pivot['calc_prod_d1'] = df_pivot['240FI020A.PNT']
    df_pivot['calc_prod_p'] = df_pivot['240FI020A.PNT']

    # unpivot
    df_unpivot = pd.melt(df_pivot.reset_index(), id_vars = ['datetime'])
    df_unpivot.columns = ['datetime', 'Tag', 'PV']
    
    return df_unpivot

preprocessed_data = crear_tags_produccion_cada_etapa(preprocessed_data)

#### 6.2 Aplicar timegap promedio
- Se necesita archivo: Maestro Etapas que tiene las etapas y sus tiempos de retención promedio
- Se necesita archivo: Maestro Tags que tiene la clasificación de los tags en cada una de las etapas

# auxiliar
def get_list_stages_to_calculate_timegap(stage_start, MaestroEtapas):
    '''
    - Definir una lista de los stages que se van a considerar para aplicar timegap.
    - El input contiene "stage_start" que indica desde qué stage se quiere considerar. Por ejemplo, si fuera D1, quiero solo
    considerar los stage desde el D1 en adelante (incluyendolo).
    '''
    
    # listado completo de stages desde el maestro etapas
    list_all_stages = MaestroEtapas['ETAPA'].tolist()
    
    # obtener solo el listado de stages desde la etapa que me interesa
    list_stages = []
    mark = False # marca para indicar que se encontró la etapa desde la que se va a considerar timegap
    for stage in list_all_stages:

        # si se llega al stage desde que se quiere aplicar el timegap hacia adelante, la marca cambia a True
        if stage == stage_start:
            mark = True

        # siolo si la marca es True, se considera guardar el stage en el listado de stages que se van a considerar
        if mark == True:
            list_stages.append(stage)
            
    return list_stages

def apply_timegap_mean(df):
    '''
    Aplicar el timegap Promedio a toda la data DESDE LA ETAPA ÁCIDA HACIA ADELANTE
    '''
    
    """ pivotear data """
    df_pivot = df.pivot(index='datetime', columns='Tag', values='PV')

    """ read MaestroTags """
    path_MaestroTags = f'config/MaestroTagsGlobal.xlsx'
    MaestroTags = pd.read_excel(path_MaestroTags)

    """ read MaestroEtapas """
    path_MaestroEtapas = f'config/MaestroEtapasGlobal.xlsx'
    MaestroEtapas = pd.read_excel(path_MaestroEtapas)

    # obtener listado de stages para timegap
    #list_stages_to_timegap = get_list_stages_to_calculate_timegap(stage_start = 'A', MaestroEtapas = MaestroEtapas)
    list_stages_to_timegap = MaestroEtapas['ETAPA'].tolist()

    """
    validar que todos los datos del dataframe estén en el maestro tags (asegurarse que todos los tags tiene asignado una etapa)
    y filtrar maestro tags tenga solo los tags presentes en la data
    """
    list_tags_data = df_pivot.columns.tolist()
    list_tags_maestro_tags = MaestroTags['TAG'].tolist()
    tags_not_in_maestro_tags = [tag for tag in list_tags_data if tag not in list_tags_maestro_tags]
    if tags_not_in_maestro_tags == []:
        print('Todos los datos del dataframe están en el maestro tags. Por lo tanto todos los tags tienen asignada una etapa')
        MaestroTags = MaestroTags.query(f'TAG == {list_tags_data}')

    """
    mover la data de acuerdo a los tiempos de residencia promedio
    """

    # inicializar
    timegap_acc = 0
    df_final = pd.DataFrame(index = df_pivot.index)

    # para cada uno de los stages...
    for stage in list_stages_to_timegap:
        print('\n\n')
        print('CALCULANDO TIEMPOS DE RESIDENCIA ETAPA:', stage)
        print('Timegap acumulado (observaciones): ', timegap_acc)


        ############## DEFINIR LISTADO DE TAGS ##############
        # filtrar el maestro tags solo por los tags que están en la etapa que se está calculando
        list_tags_stage = MaestroTags[MaestroTags['ETAPA'] == stage]['TAG'].tolist()
        print('Listado de tags del stage: ', list_tags_stage)

        # listado con el sufijo "_visto" corresponden al valor del tag visto en la realidad
        names_tag_visto = [tag + '_visto' for tag in df_pivot[list_tags_stage].columns]

        # listado con el sufijo "_real" corresponden al valor del tag desplazado de acuerdo al tiempo de residencia
        names_tag_real = [tag + '_real' for tag in df_pivot[list_tags_stage].columns]


        ############## CREAR DATASET VISTO (DATA EN EL MOMENTO ACTUAL) Y DATASET REAL (DATA DESPLAZADA DE ACUERDO AL TIMEGAP) ##############
        # data vista agregarle prefijo "visto"
        df_pivot_visto = df_pivot[list_tags_stage].copy()
        df_pivot_visto.columns = names_tag_visto

        # data real agregarle prefijo "real"
        df_pivot_real = df_pivot[list_tags_stage].copy()
        df_pivot_real.columns = names_tag_real

        # data real desplazarla de acuerdo al tiempo de residencia
        df_pivot_real = df_pivot_real.shift(-timegap_acc)


        ############## DATASET VISTO Y REAL GUARDARLOS Y TERMINAR ITERACIÓN DEL STAGE ACTUAL ##############
        # agregar data real y data vista al dataframe final
        #df_final = df_final.merge(df_pivot_visto, left_index = True, right_index = True) # TODO: COMMENT LA DATA VISTA EN EL MOMENTO ACTUAL
        df_final = df_final.merge(df_pivot_real, left_index = True, right_index = True)

        # calcular el valor de tiempo de residencia acumulado al terminar esta stage - en base a observaciones de 5 minutos
        timegap_acc += int(MaestroEtapas[MaestroEtapas['ETAPA'] == stage]['TIEMPO_RESIDENCIA_PROMEDIO'].values[0] / 5)


    df_final_unpivot = pd.melt(df_final.reset_index(), id_vars = ['datetime'])
    df_final_unpivot.columns = ['datetime', 'Tag', 'PV']
    return df_final_unpivot

preprocessed_data = apply_timegap_mean(preprocessed_data)

### 7. Pivotear data

class Pivot(BaseEstimator,TransformerMixin):
    '''
    A class used to transform the data with tags in vertical to data with tags horizontal(in columns)
    
    Input:
        original_datetime	ETAPA	Tag	            PV	        datetime
    0	2021-10-01 00:20:00	EOP	    240AIC126.MEAS	11.199728	2021-10-01 13:55:00
    1	2021-10-01 00:25:00	EOP	    240AIC126.MEAS	11.210860	2021-10-01 14:00:00
    2	2021-10-01 00:30:00	EOP	    240AIC126.MEAS	11.217164	2021-10-01 14:05:00
    
    Si se pivotea la data original (original_datetime), no existen valores nulos
    Si se pivotea la data aplicado el timegap, si existen valores nulos
    '''
    def __init__(self):
        super(Pivot,self).__init__()

    def fit(self,DataFrame):
        return self

    def transform(self,DataFrame):
        print('\naplicando pivot...')
        
        # hacer trnasformacion
        DataFrame = DataFrame.pivot_table(index = "datetime", columns = "Tag", values = "PV")
        
        #info
        print('\n Tamaño dataframe: ', DataFrame.shape)
        print('\nCantidad de nulos (estos nulos se generan a aplicar el timegap): ', DataFrame.isnull().sum())
        print('\n% de nulos (estos nulos se generan a aplicar el timegap): ', 100 * (DataFrame.isnull().sum() / DataFrame.shape[0]))
        return DataFrame

# instancia de la clase
pivoter = Pivot()

# transformar data
preprocessed_data = pivoter.transform(preprocessed_data)

### 8. Eliminar datos de PGP

class DropMaintenanceDates(BaseEstimator,TransformerMixin):
    '''
    A class to delete the dates when the plant was not operative
    '''
    
    def __init__(self,MaintenanceDataFrame,beginDateCol,endDateCol,**kwargs):
        super(DropMaintenanceDates,self).__init__()
        self.MaintenanceDataFrame = MaintenanceDataFrame 
        self.beginDateCol = beginDateCol
        self.endDateCol = endDateCol
        self.kwargs = kwargs

    def fit(self,DataFrame):
        return self

    def transform(self,DataFrame):
        print('\naplicando limpiezas fechas de mantención - PGP ...')
        print('Tamaño data: ', DataFrame.shape)
        
        for i, row in self.MaintenanceDataFrame.iterrows():
            DropIndex=DataFrame.loc[row[self.beginDateCol]:row[self.endDateCol]].index
            DataFrame.drop(DropIndex,inplace=True)
            
        #info
        print('Tamaño data luego de borrar fechas de matención: ', DataFrame.shape)
        print('\n% de nulos hasta el momento: ', 100 * (DataFrame.isnull().sum() / DataFrame.shape[0]))
        return DataFrame

def create_maintenance_df(beginning_date, ending_date):
    maintenance_df = pd.DataFrame()
    maintenance_df["beginDate"] = beginning_date
    maintenance_df["endDate"] = ending_date
    return maintenance_df

# parámetros
tag_classification_pars = load_all_parameterstags_tagclassification(model_name = general_params_models)
beginning_date = tag_classification_pars['maintenance_dates']['beginDate']
ending_date =  tag_classification_pars['maintenance_dates']['endDate']
maintenance_df = create_maintenance_df(beginning_date, ending_date)

pars_drop_maintenance_dates = {
        "MaintenanceDataFrame": maintenance_df,
        "beginDateCol": "beginDate",
        "endDateCol": "endDate",
    }

# instancia de la clase
droper_maintenance_dates = DropMaintenanceDates(**pars_drop_maintenance_dates)

# transformar dataframe
processed_data = droper_maintenance_dates.transform(preprocessed_data)

processed_data.head()

### 9. Rename columns timegap "xxx_real"
Eliminar el sufijo generado al crear el timegap "xxx_real"

# generate list without columns
list_columns = processed_data.columns
list_columns_without_suffix = [elemento.replace('_real', '') for elemento in list_columns]

# rename
processed_data.columns = list_columns_without_suffix

processed_data.head()

### 10. GUARDAR PKL PROCESSED

# save data pkl cloud
path_raw_data_processed = 'artifacts/data/data_raw_processed.pkl'
with open(path_raw_data_processed, "wb") as output:
    pickle.dump(processed_data, output)
    output.close()
